requirements.txt
# UI & Core Utilities
streamlit
pandas
python-dotenv

# LangChain Ecosystem
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
langchain-classic

# Vector Database Infrastructure
qdrant-client
langchain-qdrant

# Observability Control Plane
langfuse

# Document Ingestion Parsers
pypdf
docx2txt
openpyxl
